More clean up

+++ src/ie/gmit/sw/DomainFrequency.java
@@ -6,7 +6,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class DomainFrequency {
-    private Map<String, Integer> domainVisits;
+    private Map<String, Double> domainVisits;
     private int totalVisits;
 
     public DomainFrequency() {
@@ -17,7 +17,7 @@ public class DomainFrequency {
     public synchronized void recordVisit(String url) {
         String domain = getDomainName(url);
         if (!domainVisits.containsKey(domain)) {
-            domainVisits.put(domain, 0);
+            domainVisits.put(domain, 0d);
         }
         domainVisits.put(domain, domainVisits.get(domain) + 1);
         totalVisits++;
@@ -32,10 +32,10 @@ public class DomainFrequency {
         }
 
         // return relative frequency (value between 0 and 1)
-        return (double) domainVisits.get(domain) / totalVisits;
+        return domainVisits.get(domain) / totalVisits;
     }
 
-    public Map<String, Integer> getVisitMap() {
+    public Map<String, Double> getVisitMap() {
         return domainVisits;
     }
